requirements.txt
keras
tensorflow-gpu
flask
boto3
pandas
numpy
requests
matplotlib
opencv-python
glob2
pillow
seaborn
plotly
scikit-learn
path